+++ SyncDemo/app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
             android:name=".DocsActivity"
             android:label="@string/title_activity_docs"></activity>
         <activity
-            android:name=".sync.DiffSyncMainActivity"
+            android:name=".DiffSyncMainActivity"
             android:label="@string/title_activity_docs">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
