+++ SyncDemo/app/src/main/java/org/jboss/aerogear/android/cookbook/syncdemo/util/JsonUtil.java
@@ -1,10 +1,11 @@
-package org.jboss.aerogear.android.cookbook.syncdemo.sync;
+package org.jboss.aerogear.android.cookbook.syncdemo.util;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
+import org.jboss.aerogear.android.cookbook.syncdemo.vo.Info;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
